Moved logic back into BokehJSONEncoder
